Added attr() to Element and recusive iterators/aggregators on ElementList
The various aggregators and iterators are simple methods which take a
callback and then "map" the results in a certain way.  For aggregate() it will
merge all return values of the function into an array and return that.  For
iteratorNew it will provide a new ElementList which you can work with in the
callback and then return that ElementList itself (useful for filtered merging)
and currently used on recursive find.